chore: 컴포넌트 정리 및 폴더 컨벤션 케밥케이스로 변경(#75)
- 컴포넌트 정리 및 폴더 컨벤션 케밥케이스로 변경
- 절대경로 @custom-types로 변경
