--- src/components/Icon.tsx
@@ -1,21 +1,0 @@
-import { type IconProps } from '@customType/icon'
-import { cn } from '@utils/cn'
-
-export default function Icon({
-  icon: ImgIcon,
-  size = 18,
-  color = 'text-primary-600',
-  className,
-  iconClassName,
-}: IconProps) {
-  return (
-    <div className={cn(`flex items-center justify-center`, className)} style={{
-          width: `${size}px`,
-          height:`${size}px`,
-        }}>
-      <ImgIcon
-        className={cn(color, iconClassName)}
-      />
-    </div>
-  )
-}
