+++ src/components/common/Button.tsx
@@ -1,4 +1,7 @@
-import { BUTTON_VARIANTS, DEFAULT_BUTTON_VARIANT } from '@constants/button'
+import {
+  BUTTON_VARIANTS,
+  DEFAULT_BUTTON_VARIANT,
+} from '@constants/button/button'
 import { cn } from '@utils/cn'
 
 type ButtonVariant = keyof typeof BUTTON_VARIANTS
@@ -23,9 +26,7 @@ const Button: React.FC<ButtonProps> = ({
       onClick={onClick}
       disabled={disabled}
       className={cn(
-        'h-[36px] w-[55px] rounded-[3px] px-4 py-2',
-        'font-normal',
-        'cursor-pointer',
+        'h-[36px] w-[55px] px-4 py-2 font-semibold',
         BUTTON_VARIANTS[variant],
         { 'cursor-not-allowed': disabled },
         className
