+++ src/components/common/data-table/TableRow.tsx
@@ -1,4 +1,4 @@
-import type { TableHeader, TableRowData } from '@customType/table'
+import type { TableHeader, TableRowData } from '@custom-types/table'
 import { cn } from '@utils/cn'
 import { useState } from 'react'
