+++ src/components/layout/Layout.tsx
@@ -1,13 +1,13 @@
-import Header from '@components/Header'
-import { menuData } from '@constants/menuData'
-import SideBarMenu from '@components/Sidebar/SideBarMenu'
+import Header from '@components/layout/Header'
+import { menuData } from '@constants/layout/sidebar/menuData'
+import SideBarMenu from '@components/layout/sidebar/SideBarMenu'
 import { Outlet } from 'react-router'
 
 const Layout = () => {
   return (
     <div className="flex h-screen">
       <SideBarMenu menuData={menuData} />
-      <div className="flex flex-col flex-1">
+      <div className="flex flex-1 flex-col">
         <Header />
         <main className="flex-1">
           <Outlet />
