+++ vite.config.ts
@@ -13,7 +13,7 @@ export default defineConfig({
       '@hooks': resolve(__dirname, 'src/hooks'),
       '@lib': resolve(__dirname, 'src/lib'),
       '@pages': resolve(__dirname, 'src/pages'),
-      '@customType': resolve(__dirname, 'src/types'),
+      '@custom-types': resolve(__dirname, 'src/types'),
       '@utils': resolve(__dirname, 'src/utils'),
       '@constants': resolve(__dirname, 'src/constants'),
       '@store': resolve(__dirname, 'src/store'),
